requirements.txt
pydantic==2.7.4
typer==0.12.3
rich==13.7.1
pyyaml==6.0.1

#dev
pytest